[AIP-21] Move sensors.s3_prefix_sensor airflow.sensors.s3_prefix_sensor
