+++ scripts/ci/pylint_todo.txt
@@ -176,7 +176,7 @@
 ./airflow/providers/apache/hive/sensors/metastore_partition.py
 ./airflow/providers/apache/hive/sensors/named_hive_partition.py
 ./airflow/providers/amazon/aws/sensors/s3_key.py
-./airflow/sensors/s3_prefix_sensor.py
+./airflow/providers/amazon/aws/sensors/s3_prefix.py
 ./airflow/sensors/sql_sensor.py
 ./airflow/sensors/time_delta_sensor.py
 ./airflow/sensors/time_sensor.py
